SGN_FMSP_APP/requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
fonttools==4.52.1
fpdf2==2.7.9
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
